refactor: 조회기간, 정렬기준 컴포넌트 분리
