--- src/containers/home/common/TopBar/buttons/SearchButton/PaymentHistoryModal/InputForm.tsx
@@ -1,355 +1,0 @@
-import {
-  Stack,
-  Select,
-  InputAdornment,
-  MenuItem,
-  OutlinedInput,
-  TextField,
-  Box,
-  Grid,
-  Button,
-} from "@mui/material";
-import {
-  LocalizationProvider,
-  MobileDatePicker,
-  PickersDay,
-} from "@mui/x-date-pickers";
-import CalendarTodayOutlinedIcon from "@mui/icons-material/CalendarTodayOutlined";
-import { AdapterMoment } from "@mui/x-date-pickers/AdapterMoment";
-import moment from "moment";
-import { RenderDayFunction } from "../../../../../../../types/common";
-import { useEffect, useState } from "react";
-import {
-  BANK_ORGANIZATION,
-  CARD_ORGANIZATION,
-} from "../../../../../../../constants/organizations";
-import { useRecoilValue } from "recoil";
-import { bottomTabMenuRepository } from "@app/recoil/bottomTabMenu";
-import useHeader from "@hooks/useHeader";
-import { HEADER_MODE } from "@app/recoil/header";
-
-interface InputFormProps {
-  selected: string;
-  form: {
-    organization: string;
-    cardNo: string;
-    account: string;
-    connectedId: string;
-    startDate: string;
-    endDate: string;
-    orderBy: string;
-    inquiryType: string;
-  };
-  changeDetailInfo: (e: {
-    target: { id: string; value: string | number };
-  }) => void;
-  changeStartAndEndDate: (date: string) => void;
-}
-
-function InputForm({
-  selected,
-  form,
-  changeDetailInfo,
-  changeStartAndEndDate,
-}: InputFormProps) {
-  const [isSelectStartDate, setIsSelectStartDate] = useState(false);
-  const { openBottomBar, closeBottomBar } = useRecoilValue(bottomTabMenuRepository);
-
-  useHeader(true, HEADER_MODE.search);
-
-  useEffect(() => {
-    closeBottomBar();
-    return (() => openBottomBar());
-  }, [])
-
-  const organizations =
-    selected === "card" ? CARD_ORGANIZATION : BANK_ORGANIZATION;
-
-  const changeDate = (date: string) => {
-    if (isSelectStartDate) {
-      if (moment(date).isAfter(form.endDate)) {
-        changeStartAndEndDate(date);
-      } else {
-        changeDetailInfo({ target: { id: "startDate", value: date } });
-      }
-    } else {
-      if (moment(date).isBefore(form.startDate)) {
-        changeStartAndEndDate(date);
-      } else {
-        changeDetailInfo({ target: { id: "endDate", value: date } });
-      }
-    }
-    setIsSelectStartDate(!isSelectStartDate);
-  };
-
-  const renderDayInPicker: RenderDayFunction = (
-    day,
-    _value,
-    DayComponentProps
-  ) => {
-    if (moment(form.startDate).isSame(form.endDate)) {
-      return <PickersDay {...DayComponentProps} />;
-    }
-    if (moment(form.startDate).isSame(day)) {
-      return (
-        <PickersDay
-          sx={{
-            borderTopRightRadius: 0,
-            borderBottomRightRadius: 0,
-            marginX: 0,
-            width: "40px",
-          }}
-          className="Mui-selected"
-          {...DayComponentProps}
-        />
-      );
-    }
-    if (moment(form.endDate).isSame(day)) {
-      return (
-        <PickersDay
-          sx={{
-            borderTopLeftRadius: 0,
-            borderBottomLeftRadius: 0,
-            marginX: 0,
-            width: "40px",
-          }}
-          className="Mui-selected"
-          {...DayComponentProps}
-        />
-      );
-    }
-    if (
-      moment(form.startDate).isBefore(day) &&
-      moment(form.endDate).isAfter(day)
-    ) {
-      return (
-        <PickersDay
-          sx={{
-            borderRadius: 0,
-            marginX: 0,
-            width: "40px",
-          }}
-          className="Mui-selected"
-          {...DayComponentProps}
-        />
-      );
-    }
-    return <PickersDay {...DayComponentProps} />;
-  };
-
-  return (
-    <Stack spacing={1}>
-      <Select
-        inputProps={{
-          IconComponent: () => null,
-          style: { textAlign: "right" },
-        }}
-        size="small"
-        startAdornment={
-          <InputAdornment position="start">
-            {selected === "card" ? "카드사" : "거래은행"}
-          </InputAdornment>
-        }
-        sx={{
-          ".MuiSelect-select.MuiSelect-outlined": {
-            textAlign: "right",
-            paddingRight: "14px",
-          },
-        }}
-        value={form.organization}
-        onChange={(e) =>
-          changeDetailInfo({
-            target: { id: "organization", value: e.target.value },
-          })
-        }
-      >
-        {organizations.map((organization) => (
-          <MenuItem key={Math.random()} value={organization.value}>
-            {organization.name}
-          </MenuItem>
-        ))}
-      </Select>
-
-      <OutlinedInput
-        id={selected === "card" ? "cardNo" : "account"}
-        startAdornment={
-          <InputAdornment position="start">
-            {selected === "card" ? "카드번호" : "계좌번호"}
-          </InputAdornment>
-        }
-        value={selected === "card" ? form.cardNo : form.account}
-        onChange={changeDetailInfo}
-        size="small"
-        inputProps={{
-          style: { textAlign: "right" },
-        }}
-      />
-
-      <OutlinedInput
-        id="connectedId"
-        startAdornment={
-          <InputAdornment position="start">커넥티드ID</InputAdornment>
-        }
-        value={form.connectedId}
-        onChange={changeDetailInfo}
-        size="small"
-        inputProps={{
-          style: { textAlign: "right" },
-        }}
-      />
-
-      {/* <LocalizationProvider dateAdapter={AdapterMoment}>
-        <MobileDatePicker
-          value={moment(form.startDate)}
-          onChange={(newValue) => {
-            newValue && changeDate(newValue.format("YYYY-MM-DD"));
-          }}
-          ToolbarComponent={() => (
-            <Stack direction="row" spacing={1} justifyContent="center" p={2}>
-              <Box sx={{ color: isSelectStartDate ? "grey" : "black" }}>
-                {form.startDate}
-              </Box>
-              <Box>~</Box>
-              <Box sx={{ color: isSelectStartDate ? "black" : "grey" }}>
-                {form.endDate}
-              </Box>
-            </Stack>
-          )}
-          renderDay={renderDayInPicker}
-          componentsProps={{
-            actionBar: {
-              actions: ["accept"],
-            },
-          }}
-          renderInput={(params) => (
-            <TextField
-              {...params}
-              InputProps={{
-                startAdornment: (
-                  <InputAdornment position="start">조회일자</InputAdornment>
-                ),
-                endAdornment: (
-                  <InputAdornment position="end">
-                    <CalendarTodayOutlinedIcon
-                      fontSize="small"
-                      color={form.endDate === "" ? "secondary" : "primary"}
-                    />
-                  </InputAdornment>
-                ),
-              }}
-              inputProps={{
-                style: { textAlign: "right" },
-              }}
-              size="small"
-              value={`${form.startDate}~${form.endDate}`}
-            />
-          )}
-        />
-      </LocalizationProvider> */}
-
-      {/* {selected === "card" && (
-        <Select
-          inputProps={{
-            IconComponent: () => null,
-            style: { textAlign: "right" },
-          }}
-          size="small"
-          startAdornment={
-            <InputAdornment position="start">조회구분</InputAdornment>
-          }
-          sx={{
-            ".MuiSelect-select.MuiSelect-outlined": {
-              textAlign: "right",
-              paddingRight: "14px",
-            },
-          }}
-          id="inquiryType"
-          value={form.inquiryType}
-          onChange={(e) =>
-            changeDetailInfo({
-              target: { id: "inquiryType", value: e.target.value },
-            })
-          }
-        >
-          <MenuItem value={"0"}>카드별조회</MenuItem>
-          <MenuItem value={"1"}>전체조회</MenuItem>
-        </Select>
-      )} */}
-      <Grid container>
-        <Grid xs="auto" item sx={{ paddingRight: "8px" }}>
-          <Box sx={{ textAlign: 'center', padding: '8px 12px', border: '1.4px solid var(--main-01, #735BF2)', borderRadius: '4px', fontSize: '14px' }}>조회기간</Box>
-        </Grid>
-        <Grid xs item>
-          <LocalizationProvider dateAdapter={AdapterMoment}>
-            <MobileDatePicker
-              value={moment(form.startDate)}
-              onChange={(newValue) => {
-                newValue && changeDate(newValue.format("YYYY-MM-DD"));
-              }}
-              ToolbarComponent={() => (
-                <Stack direction="row" spacing={1} justifyContent="center" p={2}>
-                  <Box sx={{ color: isSelectStartDate ? "grey" : "black" }}>
-                    {form.startDate}
-                  </Box>
-                  <Box>~</Box>
-                  <Box sx={{ color: isSelectStartDate ? "black" : "grey" }}>
-                    {form.endDate}
-                  </Box>
-                </Stack>
-              )}
-              renderDay={renderDayInPicker}
-              componentsProps={{
-                actionBar: {
-                  actions: ["accept"],
-                },
-              }}
-              renderInput={(params) => (
-                <TextField
-                  {...params}
-                  fullWidth
-                  InputProps={{
-                    endAdornment: (
-                      <InputAdornment position="end">
-                        <CalendarTodayOutlinedIcon
-                          fontSize="small"
-                          color={form.endDate === "" ? "secondary" : "primary"}
-                        />
-                      </InputAdornment>
-                    ),
-                  }}
-                  inputProps={{
-                    style: { textAlign: "right" },
-                  }}
-                  size="small"
-                  value={`${form.startDate}~${form.endDate}`}
-                />
-              )}
-            />
-          </LocalizationProvider>
-        </Grid>
-      </Grid>
-
-      <Grid container>
-        <Grid xs="auto" item sx={{ paddingRight: "8px" }}>
-          <Box sx={{ textAlign: 'center', padding: '8px 12px', border: '1.4px solid var(--main-01, #735BF2)', borderRadius: '4px', fontSize: '14px' }}>정렬기준</Box>
-        </Grid>
-        <Grid xs item sx={{ paddingRight: "8px" }}>
-          <Button fullWidth variant="contained" sx={{ borderRadius: "4px 4px 0px 0px" }} onChange={(e) =>
-            changeDetailInfo({
-              target: { id: "inquiryType", value: "0" },
-            })
-          }>최신순</Button>
-        </Grid>
-        <Grid xs item>
-          <Button fullWidth variant="outlined" sx={{ borderRadius: "4px 4px 0px 0px", borderColor: "#A9ACB2", color: "#5B5F67" }} onChange={(e) =>
-            changeDetailInfo({
-              target: { id: "inquiryType", value: "1" },
-            })
-          }>과거순</Button>
-        </Grid>
-      </Grid>
-    </Stack>
-  );
-}
-
-export default InputForm;
